moved random front facing fcns to lc
