+++ localization/vision_common/include/vision_common/test_utilities.h
@@ -55,16 +55,6 @@ Eigen::VectorXd RandomRadDistortion();
 
 Eigen::VectorXd RandomRadTanDistortion();
 
-// Samples in cylindrical coordinates for pose translation to keep pose in view frustrum.
-// Samples z using scaled rho value to prevent large z vals with small rho values
-// that may move the pose out of the view frustrum.
-Eigen::Isometry3d RandomFrontFacingPose(const double rho_min, const double rho_max, const double phi_min,
-                                        const double phi_max, const double z_rho_scale, const double yaw_min,
-                                        const double yaw_max, const double pitch_min, const double pitch_max,
-                                        const double roll_min, const double roll_max);
-
-Eigen::Isometry3d RandomFrontFacingPose();
-
 // Spaced out poses for targets which when projected into image space cover
 // the image well with target points.  Poses are sampled for each row/col combination
 // and evenly spaced in cylindrical coordinates
@@ -73,13 +63,6 @@ std::vector<Eigen::Isometry3d> EvenlySpacedTargetPoses(const int num_rows = 3, c
 
 std::vector<Eigen::Vector3d> TargetPoints(const int points_per_row, const int points_per_col,
                                           const double row_spacing = 0.1, const double col_spacing = 0.1);
-
-std::vector<Eigen::Vector3d> RandomFrontFacingPoints(const int num_points);
-
-Eigen::Vector3d RandomFrontFacingPoint();
-
-Eigen::Vector3d RandomFrontFacingPoint(const double rho_min, const double rho_max, const double phi_min,
-                                       const double phi_max, const double z_rho_scale);
 
 template <typename DISTORTER>
 class RegistrationCorrespondences {
