+++ localization/vision_common/src/test_utilities.cc
@@ -141,65 +141,6 @@ std::vector<Eigen::Vector3d> TargetPoints(const int points_per_row, const int po
   return target_points;
 }
 
-std::vector<Eigen::Vector3d> RandomFrontFacingPoints(const int num_points) {
-  std::vector<Eigen::Vector3d> points;
-  for (int i = 0; i < num_points; ++i) {
-    points.emplace_back(RandomFrontFacingPoint());
-  }
-  return points;
-}
-
-Eigen::Vector3d RandomFrontFacingPoint() {
-  static constexpr double rho_min = 1.0;
-  static constexpr double rho_max = 3.0;
-  static constexpr double phi_min = -25.0;
-  static constexpr double phi_max = 25.0;
-  static constexpr double z_rho_scale = 0.5;
-  return RandomFrontFacingPoint(rho_min, rho_max, phi_min, phi_max, z_rho_scale);
-}
-
-Eigen::Vector3d RandomFrontFacingPoint(const double rho_min, const double rho_max, const double phi_min,
-                                       const double phi_max, const double z_rho_scale) {
-  const double rho = lc::RandomDouble(rho_min, rho_max);
-  const double phi = lc::RandomDouble(phi_min, phi_max);
-  const double z = lc::RandomDouble(-1.0 * z_rho_scale * rho, z_rho_scale * rho);
-  // Z and x are swapped so z defines distance from camera rather than height
-  const Eigen::Vector3d tmp = lc::CylindricalToCartesian(Eigen::Vector3d(rho, phi, z));
-  const Eigen::Vector3d random_point(tmp.z(), tmp.y(), tmp.x());
-  return random_point;
-}
-
-Eigen::Isometry3d RandomFrontFacingPose() {
-  static constexpr double rho_min = 1.0;
-  static constexpr double rho_max = 3.0;
-  static constexpr double phi_min = -25.0;
-  static constexpr double phi_max = 25.0;
-  static constexpr double z_rho_scale = 0.5;
-
-  // Pitch acts like yaw since z axis points outwards in camera frame
-  static constexpr double yaw_min = -10.0;
-  static constexpr double yaw_max = 10.0;
-  static constexpr double pitch_min = -45;
-  static constexpr double pitch_max = 45;
-  static constexpr double roll_min = -10;
-  static constexpr double roll_max = 10;
-
-  return RandomFrontFacingPose(rho_min, rho_max, phi_min, phi_max, z_rho_scale, yaw_min, yaw_max, pitch_min, pitch_max,
-                               roll_min, roll_max);
-}
-
-Eigen::Isometry3d RandomFrontFacingPose(const double rho_min, const double rho_max, const double phi_min,
-                                        const double phi_max, const double z_rho_scale, const double yaw_min,
-                                        const double yaw_max, const double pitch_min, const double pitch_max,
-                                        const double roll_min, const double roll_max) {
-  const Eigen::Vector3d translation = RandomFrontFacingPoint(rho_min, rho_max, phi_min, phi_max, z_rho_scale);
-  const double yaw = lc::RandomDouble(yaw_min, yaw_max);
-  const double pitch = lc::RandomDouble(pitch_min, pitch_max);
-  const double roll = lc::RandomDouble(roll_min, roll_max);
-  const Eigen::Matrix3d rotation = lc::RotationFromEulerAngles(yaw, pitch, roll);
-  return lc::Isometry3d(translation, rotation);
-}
-
 std::vector<Eigen::Isometry3d> EvenlySpacedTargetPoses(const int num_rows, const int num_cols, const int num_y_levels) {
   // Pitch acts like yaw since z axis points out of camera frame
   const lc::Sampler pitch_sampler(-15.0, 15.0, num_cols);
